+++ rasa_core/channels/console.py
@@ -34,6 +34,11 @@ def print_bot_output(message, color=utils.bcolors.OKBLUE):
             element_str = "Elements:\n" + element_to_string(element, idx)
             utils.print_color(element_str, color)
 
+    if "quick_replies" in message:
+        for idx, element in enumerate(message.get("quick_replies")):
+            element_str = "Quick Replies:\n" + button_to_string(element, idx)
+            utils.print_color(element_str, color)
+
 
 def get_cmd_input():
     response = questionary.text("",