+++ rasa_core/channels/facebook.py
@@ -161,7 +161,18 @@ class MessengerBot(OutputChannel):
                                        {"sender": {"id": recipient_id}},
                                        'RESPONSE')
 
-    def send_custom_message(self, recipient_id: Text,
+    def send_quick_replies(self,
+                           recipient_id: Text,
+                           text: Text,
+                           quick_replies: List[Dict[Text, Any]],
+                           **kwargs: Any) -> None:
+        """Sends quick replies to the output."""
+
+        self._add_text_info(quick_replies)
+        self.send(recipient_id, FBText(text=text, quick_replies=quick_replies))
+
+    def send_custom_message(self,
+                            recipient_id: Text,
                             elements: List[Dict[Text, Any]]) -> None:
         """Sends elements to the output."""
 
@@ -183,10 +194,19 @@ class MessengerBot(OutputChannel):
 
     @staticmethod
     def _add_postback_info(buttons: List[Dict[Text, Any]]) -> None:
-        """Make sure every button has a type. Modifications happen in place."""
+        """Set the button type to postback for all buttons without type.
+        Happens in place."""
         for button in buttons:
-            if 'type' not in button:
+            if not button.get('type'):
                 button['type'] = "postback"
+
+    @staticmethod
+    def _add_text_info(quick_replies: List[Dict[Text, Any]]) -> None:
+        """Set the quick reply type to text for all buttons without content
+        type. Happens in place."""
+        for quick_reply in quick_replies:
+            if not quick_reply.get('type'):
+                quick_reply['content_type'] = "text"
 
     @staticmethod
     def _recipient_json(recipient_id: Text) -> Dict[Text, Dict[Text, Text]]: